LPDEV-151 double to decimal

+++ src/Lykke.Service.PayInternal.AzureRepositories/PaymentRequest/PaymentRequestEntity.cs
@@ -19,7 +19,7 @@ namespace Lykke.Service.PayInternal.AzureRepositories.PaymentRequest
         public string Id => RowKey;
         public string MerchantId { get; set; }
         public string OrderId { get; set; }
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
         public string SettlementAssetId { get; set; }
         public string PaymentAssetId { get; set; }
         public DateTime DueDate { get; set; }
@@ -27,7 +27,7 @@ namespace Lykke.Service.PayInternal.AzureRepositories.PaymentRequest
         public int MarkupPips { get; set; }
         public string WalletAddress { get; set; }
         public PaymentRequestStatus Status { get; set; }
-        public double PaidAmount { get; set; }
+        public decimal PaidAmount { get; set; }
         public DateTime? PaidDate { get; set; }
         public string Error { get; set; }
