+++ src/Lykke.Service.PayInternal.Contract/PaymentRequest/PaymentRequestDetailsMessage.cs
@@ -16,7 +16,7 @@ namespace Lykke.Service.PayInternal.Contract.PaymentRequest
         
         public string MerchantId { get; set; }
         
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
         
         public string SettlementAssetId { get; set; }
         
@@ -33,7 +33,7 @@ namespace Lykke.Service.PayInternal.Contract.PaymentRequest
         [JsonConverter(typeof(StringEnumConverter))]
         public PaymentRequestStatus Status { get; set; }
         
-        public double PaidAmount { get; set; }
+        public decimal PaidAmount { get; set; }
         
         public DateTime? PaidDate { get; set; }
