+++ src/Lykke.Service.PayInternal.Services/Domain/PaymentRequest.cs
@@ -8,7 +8,7 @@ namespace Lykke.Service.PayInternal.Services.Domain
         public string Id { get; set; }
         public string MerchantId { get; set; }
         public string OrderId { get; set; }
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
         public string SettlementAssetId { get; set; }
         public string PaymentAssetId { get; set; }
         public DateTime DueDate { get; set; }
@@ -16,7 +16,7 @@ namespace Lykke.Service.PayInternal.Services.Domain
         public int MarkupPips { get; set; }
         public string WalletAddress { get; set; }
         public PaymentRequestStatus Status { get; set; }
-        public double PaidAmount { get; set; }
+        public decimal PaidAmount { get; set; }
         public DateTime? PaidDate { get; set; }
         public string Error { get; set; }
     }
